SnacSequence is DWORD so we should treat it as DWORD and not cast it to WORD as we did in some places because then we can end up comparing DWORD and WORD. The internal SnacSequence counter is still WORD as I don't know if we can use full DWORD.
Fixes bug when sending animation never stops.


svn path=/trunk/KDE/kdenetwork/kopete/; revision=879289

+++ protocols/oscar/liboscar/transfer.h
@@ -161,7 +161,7 @@ private:
 	Oscar::WORD m_snacService;
 	Oscar::WORD m_snacSubtype;
 	Oscar::WORD m_snacFlags;
-	Oscar::WORD m_snacReqId;
+	Oscar::DWORD m_snacReqId;
 	
 	bool m_isSnacValid;
 };
